[utils] Add scripting.py
* Update snapshot_recover.py to newly added utils
